pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bank-check-processor"
version = "1.0.0"
description = "Automated Bank Check Processing System"
authors = [{name = "Bank Check Processor", email = "admin@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "streamlit>=1.29.0",
    "google-generativeai>=0.3.2",
    "Pillow>=10.1.0",
    "pandas>=2.1.4",
    "plotly>=5.17.0",
    "opencv-python-headless==4.8.1.78",
    "flask==2.3.3",
    "numpy>=1.25.2"
]

[project.scripts]
start = "streamlit run app.py --server.port=$PORT --server.address=0.0.0.0"
